refactor(WebexActivity): move activity header into its own file
