+++ src/components/WebexActivity/WebexActivity.test.js
@@ -1,19 +1,10 @@
 import React from 'react';
 
-import jsonData from '../../data';
-
-import WebexActivity, {Header, formatMessageDate} from './WebexActivity';
+import WebexActivity from './WebexActivity';
 
 jest.mock('../hooks/useActivity');
-jest.mock('../hooks/usePerson');
 
 describe('Webex Activity component', () => {
-  describe('Header component snapshot', () => {
-    test('matches snapshot with "default" props', () => {
-      expect(shallow(<Header personID="default" created={jsonData.activities.default.created} />)).toMatchSnapshot();
-    });
-  });
-
   describe('Webex Activity snapshots', () => {
     test('matches snapshot with "default" text', () => {
       expect(shallow(<WebexActivity activityID="default" />)).toMatchSnapshot();
@@ -41,32 +32,6 @@ describe('Webex Activity component', () => {
 
     test('matches snapshot with a text over a week ago', () => {
       expect(shallow(<WebexActivity activityID="old" />)).toMatchSnapshot();
-    });
-  });
-
-  describe('unit testing', () => {
-    test('formatMessageDate() returns today date', () => {
-      expect(formatMessageDate(new Date('today'))).toEqual('today p');
-    });
-
-    test('formatMessageDate() returns today date', () => {
-      expect(formatMessageDate(new Date('yesterday'))).toEqual('Yesterday yesterday p');
-    });
-
-    test('formatMessageDate() returns same week date', () => {
-      expect(formatMessageDate(new Date('sameWeek'))).toEqual('sameWeek iiii p');
-    });
-
-    test('formatMessageDate() returns today date', () => {
-      expect(formatMessageDate(new Date('old'))).toEqual('old P p');
-    });
-
-    test('throws error with inappropriate activityID', () => {
-      const wrongID = 'Wrong activityID';
-
-      expect(() => shallow(<WebexActivity activityID={wrongID} />)).toThrowError(
-        new Error(`Could not find activity with ID "${wrongID}"`)
-      );
     });
   });
 });
